Username should be checked from google (because can be longer then 10 characters) @Chr-ss [DONE]
All players join and leave the tournament. TOURNAMENT_ERROR Player(s) not found { player1: undefined, player2: undefined } createTournamentMatch - tournament.js:147 if (!player1.ready || !player2.ready)
[DONE]

Create an account normally that is the same as the google login. Then login with google will redirect to the google callback, but should go to login page which is saying "User already exists - please log in normally" or something. @Chr-ss
- Message is changed to "User already exists - please log in with your username and password."
- Redirection to 500 page stays for simplicity and clarity since it is an error case.
[DONE]

Log in with google as the second user will take over the first player. So you can also not log in as two google users @Chr-ss
[DONE]

You are able to log in twice with google, but should get the notification: user already logged in. @Chr-ss
[DONE]