Username limit was only working for updates. I fixed it and tested with new accounts on P1 and P2. 	- I also fixed a bug where the Google ID was invalidated when the user Data changed which prevented further logins with out system
Login with Google accounts on P2 works. I fixed a Frontend mussing the correct User information following Merels example for the Username that was already in the main. Google Login on P1 and P2 is not overwriting the Logins and you are recognized correctly in tournaments, online games and local games.

Google Login is bound to the online state. I am not changing the redirect because we want to eval. This might have other errors I can not forsee currently.

All players join and leave the tournament. >> this works if you don't spam it with the same Users